backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1

